fixing and refactoring

+++ src/controllers/question-types.controller.js
@@ -1,27 +1,40 @@
 const QuestionTypesModel = require('../models/question-types.model');
 
-const getQuestionTypes = async (req, res) => {
+const BadRequestError = require('../errors/bad-request');
+
+const getQuestionTypes = async (req, res, next) => {
   const model = new QuestionTypesModel();
 
   try {
     const questionTypes = await model.findAll();
-    res.json({ questionTypes });
+    res.json(questionTypes);
   } catch (error) {
-    console.error(error);
-    res.status(500).send('Internal server error');
+    if (error?.message) {
+      const err = new BadRequestError(error.message);
+      return next(err);
+    } else {
+      console.error(error);
+      res.status(500).send('Internal server error');
+    }
   }
 };
 
-const getQuestionTypeById = async (req, res) => {
+const getQuestionTypeById = async (req, res, next) => {
   const model = new QuestionTypesModel();
   const questionTypeId = req.params.id;
 
+  if (isNaN(questionTypeId)) {
+    const err = new BadRequestError('Error! You need to provide valid id.');
+    return next(err);
+  }
+
   try {
     const questionType = await model.findById(questionTypeId);
-    res.json({ questionType });
+    res.json({ ...questionType });
   } catch (error) {
     if (error?.message) {
-      res.status(404).json({ error: error.message });
+      const err = new BadRequestError(error.message);
+      return next(err);
     } else {
       console.error(error);
       res.status(500).send('Internal server error');
@@ -29,15 +42,29 @@ const getQuestionTypeById = async (req, res) => {
   }
 };
 
-const createQuestionType = async (req, res) => {
+const createQuestionType = async (req, res, next) => {
   const model = new QuestionTypesModel();
 
+  if (!req.body.Type || Object.keys(req.body).length !== 1) {
+    const err = new BadRequestError('Error! You need to provide Type.');
+    return next(err);
+  }
+
   try {
-    await model.create(req.body);
+    const response = await model.findAll();
+    response.forEach((e) => {
+      if (
+        e.Type.replace(/\s+/g, '').toLowerCase() ===
+        req.body.Type.replace(/\s+/g, '').toLowerCase()
+      )
+        throw new Error('Type actully exist');
+    });
+    await model.create({ Type: req.body.Type.replace(/\s+/g, ' ') });
     res.status(201).json({ message: `Question type has been created` });
   } catch (error) {
     if (error?.message) {
-      res.status(404).json({ error: error.message });
+      const err = new BadRequestError(error.message);
+      return next(err);
     } else {
       console.error(error);
       res.status(500).send('Internal server error');
@@ -45,22 +72,29 @@ const createQuestionType = async (req, res) => {
   }
 };
 
-const updateQuestionType = async (req, res) => {
+const updateQuestionType = async (req, res, next) => {
   const model = new QuestionTypesModel();
   const questionTypeId = req.params.id;
   const questionTypeBody = req.body;
 
-  try {
-    const questionTyp = await model.findById(questionTypeId);
+  if (isNaN(questionTypeId)) {
+    const err = new BadRequestError('Error! You need to provide valid id.');
+    return next(err);
+  }
 
-    await model.update(questionTypeId, {
-      ...questionTyp[0],
-      ...questionTypeBody,
-    });
+  if (!questionTypeBody.Type) {
+    const err = new BadRequestError('Error! You need to provide Type.');
+    return next(err);
+  }
+
+  try {
+    await model.findById(questionTypeId);
+    await model.update(questionTypeId, questionTypeBody);
     res.status(201).json({ message: `Question typw has been updated` });
   } catch (error) {
     if (error?.message) {
-      res.status(404).json({ error: error.message });
+      const err = new BadRequestError(error.message);
+      return next(err);
     } else {
       console.error(error);
       res.status(500).send('Internal server error');
@@ -68,21 +102,25 @@ const updateQuestionType = async (req, res) => {
   }
 };
 
-const deleteQuestionTypeById = async (req, res) => {
+const deleteQuestionTypeById = async (req, res, next) => {
   const model = new QuestionTypesModel();
   const questionTypeId = req.params.id;
+
+  if (isNaN(questionTypeId)) {
+    const err = new BadRequestError('Error! You need to provide valid id.');
+    return next(err);
+  }
 
   try {
     await model.findById(questionTypeId);
     await model.delete(questionTypeId);
-    res
-      .status(200)
-      .json({
-        message: `Question type with ID ${questionTypeId} has been deleted`,
-      });
+    res.status(200).json({
+      message: `Question type with ID ${questionTypeId} has been deleted`,
+    });
   } catch (error) {
     if (error?.message) {
-      res.status(404).json({ error: error.message });
+      const err = new BadRequestError(error.message);
+      return next(err);
     } else {
       console.error(error);
       res.status(500).send('Internal server error');
